+++ app/app.py
@@ -6,10 +6,10 @@ import uvloop
 
 from sanic.log import logger
 
-from config_core import SEARCH_CONFIG
+from config.config_core import SEARCH_CONFIG
 
-from server.log_config import default_log_config
-from server.sanic_elasticsearch import SanicElasticsearch
+from app.log_config import default_log_config
+from app.sanic_elasticsearch import SanicElasticsearch
 
 from server.search.routes import search_blueprint
 
@@ -26,7 +26,7 @@ def create_app() -> SanicElasticsearch:
     app = SanicElasticsearch(log_config=default_log_config)
 
     logger.info("Using config '%s'" % SEARCH_CONFIG)
-    app.config.from_pyfile('config_%s.py' % SEARCH_CONFIG)
+    app.config.from_pyfile('config/config_%s.py' % SEARCH_CONFIG)
 
     # Register blueprints
     app.blueprint(search_blueprint)
