+++ app/elasticsearch_client_service.py
