+++ config/config_development.py
@@ -1,0 +1,1 @@
+TESTING = False
