+++ config/config_production.py
@@ -1,0 +1,1 @@
+TESTING = False
