+++ config/config_testing.py
@@ -1,4 +1,4 @@
-from config_core import *
+from config.config_core import *
 
 TESTING = True
 ENABLE_PROMETHEUS_METRICS = False
