--- config_production.py
@@ -1,3 +1,0 @@
-from config_core import *
-
-TESTING = False
