+++ manager.py
@@ -25,7 +25,7 @@ def run(app_host: str='0.0.0.0', app_port: int=5000, app_workers: int=1):
     :param app_workers: Number of worker threads to use (defaults to 1)
     :return:
     """
-    from server.app import create_app
+    from app.app import create_app
 
     # Create the app
     app = create_app()
