+++ manager_gunicorn.py
@@ -1,2 +1,2 @@
-from server.app import create_app
+from app.app import create_app
 app = create_app()
