+++ unit/utils/test_app.py
@@ -5,9 +5,9 @@ import unittest
 
 from urllib import parse as urllib_parse
 
-from server.app import create_app
+from app.app import create_app
 from server.request.ons_request import ONSRequest
-from server.sanic_elasticsearch import SanicElasticsearch
+from app.sanic_elasticsearch import SanicElasticsearch
 
 
 class TestApp(unittest.TestCase):
